flask/requirements.txt
flask
mysql-connector-python
uwsgi
